+++ backend/store/postgres/watcher.go
@@ -1,9 +1,7 @@
 package postgres
 
 import (
-	"database/sql"
 	"sync"
-	"time"
 
 	"github.com/jackc/pgx/v4/pgxpool"
 	"github.com/sensu/sensu-go/backend/poll"
@@ -27,25 +25,4 @@ func getWatchStoreOverride(storeName string) (factory watchStoreFactory, ok bool
 	defer watchStoreOverridesMu.Unlock()
 	factory, ok = watchStoreOverrides[storeName]
 	return
-}
-
-// recordStatus used by postgres stores implementing poll.Table
-type recordStatus struct {
-	CreatedAt time.Time
-	UpdatedAt time.Time
-	DeletedAt sql.NullTime
-}
-
-// Row builds a poll.Row from a scanned row
-func (rs recordStatus) Row(id string, resource storev2.Wrapper) poll.Row {
-	row := poll.Row{
-		Id:        id,
-		Resource:  resource,
-		CreatedAt: rs.CreatedAt,
-		UpdatedAt: rs.UpdatedAt,
-	}
-	if rs.DeletedAt.Valid {
-		row.DeletedAt = &rs.DeletedAt.Time
-	}
-	return row
 }
